refactor: extract shared OCI logic to cella-oci
