+++ crates/cella-oci/Cargo.toml
@@ -1,34 +1,21 @@
 [package]
-name = "cella-templates"
+name = "cella-oci"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
 
 [dependencies]
-cella-jsonc.workspace = true
+base64.workspace = true
 dirs.workspace = true
 flate2.workspace = true
-glob.workspace = true
-hex.workspace = true
-miette.workspace = true
 oci-distribution.workspace = true
-reqwest.workspace = true
 serde.workspace = true
 serde_json.workspace = true
-sha2.workspace = true
 tar.workspace = true
-thiserror.workspace = true
-tokio.workspace = true
 tracing.workspace = true
 
-cella-oci.workspace = true
-
 [dev-dependencies]
-filetime.workspace = true
 tempfile.workspace = true
-
-[features]
-integration-tests = []
 
 [lints]
 workspace = true
